item and owner_id connection, only approved posts are shown on browse.html now and admin panel e shob dekha jai, connection on homepage to let users get to their profile if theyre logged in home page thekei(frontend e implement korte lagbe) and rating column added for items in models.py(commented out)

+++ routers/items.py
@@ -1,4 +1,4 @@
-from fastapi import APIRouter, Request, Depends ,Form
+from fastapi import APIRouter, Request, Depends ,Form, Cookie, HTTPException
 from fastapi.responses import RedirectResponse
 from fastapi.templating import Jinja2Templates
 from sqlalchemy.orm import Session
@@ -11,14 +11,14 @@ router = APIRouter()
 templates = Jinja2Templates(directory="templates")
 
 
-@router.get("/items")
+@router.get("/items") #only approved items will show up here, the ones pending approval can only be seen on the admin panel
 def read_items(request: Request , db: Session = Depends(get_db), page: int =1, limit: int=5): #calculating how many items to show on the page
     #how many items per page
     offset = (page - 1) * limit #which item to start from on the page
     # Get the total number of items in the database 
-    total = db.query(models.Item).count()
+    total = db.query(models.Item).filter(models.Item.is_approved == True).count()
     #items on the current page
-    items = db.query(models.Item).offset(offset).limit(limit).all()
+    items = db.query(models.Item).filter(models.Item.is_approved == True).offset(offset).limit(limit).all()
     # Round up so partial pages still get their own page number
     total_pages = math.ceil(total / limit)
     return templates.TemplateResponse("browse.html", {
@@ -31,9 +31,15 @@ def read_items(request: Request , db: Session = Depends(get_db), page: int =1, l
 @router.post("/items")
 def create_item(
     form_data: ItemCreateForm = Depends(), #catching the whole form at once
+    user_email: str=Cookie(None),#user email used to make post
     db: Session = Depends(get_db)
-):
-    new_item = models.Item(name=form_data.name, description=form_data.description)
+    ):
+    
+    if not user_email:
+        return RedirectResponse(url="/login?error=You aren't logged in.", status_code=303)
+    user=db.query(models.User).filter(models.User.email==user_email).first()#matching the user to the email used to crreate the item post
+    
+    new_item = models.Item(name=form_data.name, description=form_data.description, owner_id=user.id)#last one links the post to the owner
     
     db.add(new_item)
     db.commit()
@@ -43,11 +49,21 @@ def create_item(
 @router.post("/items/delete/{item_id}")
 def delete_item(
     item_id: int , #The id that would be deleted
+    user_email: str=Cookie(None),#email of the owner
     db: Session = Depends(get_db)
 ):
+    if not user_email:
+        return RedirectResponse(url="/login?error=You aren't logged in.", status_code=303)
+    user=db.query(models.User).filter(models.User.email==user_email).first()#matching the user to the email used to crreate the item so that only they can delete the post
+    
     #Finding that item
     item= db.query(models.Item).filter(models.Item.id== item_id).first()
-
+    if not item:
+        return RedirectResponse(url="/items", status_code=303)
+    
+    if item.owner_id!=user.id and not user.is_admin:
+        raise HTTPException(status_code=403, detail="You are not authorized to delete this item") #only the owner can delete their own posts, not anyone and the admin can delete any post
+    
     if item: #checking if the item really exists
         db.delete(item)
         db.commit()
@@ -58,13 +74,18 @@ def delete_item(
 def edit_item(
     item_id: int,
     name: str = Form(...),
+    user_email: str=Cookie(None),#email of the owner
     description: str = Form(...),
     db: Session = Depends(get_db)
     ):
 
     item = db.query(models.Item).filter(models.Item.id == item_id).first()
+    if not user_email:
+        return RedirectResponse(url="/login?error=You aren't logged in.", status_code=303)
+    user=db.query(models.User).filter(models.User.email==user_email).first()#matching the user to the email used to crreate the item so that only they can delete the post
+    
 
-    if item:
+    if item and item.owner_id==user.id:
         item.name = name 
         item.description = description
         db.commit()
